requirements.txt
apache-airflow==2.10.3
kafka-python==2.2.2
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
dbt-core==1.10.4
dbt-snowflake==1.10.0
requests==2.31.0
